Refactor server pinging into shared file
